refactor(web): rename `HardKeyboard` to `HardKeyboardBase`
