+++ web/src/engine/main/src/index.ts
@@ -1,6 +1,6 @@
 export { EngineConfiguration, InitOptionDefaults, InitOptionSpec } from './engineConfiguration.js';
 export { ContextManagerBase, ContextManagerConfiguration } from './contextManagerBase.js';
-export { HardKeyboard, processForMnemonicsAndLegacy } from './hardKeyboard.js';
+export { HardKeyboardBase, processForMnemonicsAndLegacy } from './hardKeyboardBase.js';
 export { KeyboardInterfaceBase } from './keyboardInterfaceBase.js';
 export { KeymanEngineBase } from './keymanEngineBase.js';
 export { LegacyAPIEvents } from './legacyAPIEvents.js';
