+++ web/src/engine/main/src/keymanEngineBase.ts
@@ -11,7 +11,7 @@ import { ModelSpec, PredictionContext } from "keyman/engine/interfaces";
 import { EngineConfiguration, InitOptionSpec } from "./engineConfiguration.js";
 import { KeyboardInterfaceBase } from "./keyboardInterfaceBase.js";
 import { ContextManagerBase } from "./contextManagerBase.js";
-import { HardKeyboard as HardKeyboardBase } from "./hardKeyboard.js";
+import { HardKeyboardBase } from "./hardKeyboardBase.js";
 import { LegacyAPIEvents } from "./legacyAPIEvents.js";
 import { EventNames, EventListener, LegacyEventEmitter } from "keyman/engine/events";
 import DOMCloudRequester from "keyman/engine/keyboard-storage/dom-requester";
